+++ Ribbon/ColorHelper.cs
@@ -17,6 +17,13 @@ namespace RibbonLib
         public double H;
         public double S;
         public double L;
+
+        public HSL(double h, double s, double l)
+        {
+            H = h;
+            S = s;
+            L = l;
+        }
     }
 
     public struct HSB
@@ -38,8 +45,10 @@ namespace RibbonLib
         /// <returns>Ribbon Color format</returns>
         public static uint RGBToUInt32(Color color)
         {
-            HSL hsl = RGBToHSL(color);
-            HSB hsb = HSLToHSB(hsl);
+            HSL hsl = RGBToHsl(color.R, color.G, color.B);
+            //HSL hsl = RGBToHSL(color); //old one
+            //HSB hsb = HSLToHSB(hsl);
+            HSB hsb = HSLToHSB(new HSL(hsl.H / 360.0, hsl.S, hsl.L));
             uint result = HSBToUInt32(hsb);
             return result;
         }
@@ -53,10 +62,12 @@ namespace RibbonLib
         {
             HSB hsb = UInt32ToHSB(value);
             HSL hsl = HSBToHSL(hsb);
-            Color result = HSLToRGB(hsl);
+            Color result = HslToRgb(hsl.H * 360.0, hsl.S, hsl.L);
+            //Color result = HSLToRGB(hsl);
             return result;
         }
 
+        //HSB.H [0..255], HSB.S [0..255], HSB.L [0..255]
         public static HSB UInt32ToHSB(uint value)
         {
             HSB hsb = new HSB();
@@ -66,12 +77,24 @@ namespace RibbonLib
             return hsb;
         }
 
+        //HSL.H [0..1], HSL.S [0..1], HSL.L [0..1]
+        //HSB.H [0..255], HSB.S [0..255], HSB.L [0..255]
+        /// <summary>
+        /// HSB to HSL conversion. All values of HSL in Range [0..1]
+        /// </summary>
+        /// <param name="hsb"></param>
+        /// <returns></returns>
         public static HSL HSBToHSL(HSB hsb)
         {
             HSL hsl = new HSL();
             double ld;
             // Convert B to L
-            ld = Math.Exp((hsb.B - 257.7) / 149.9);
+            if (hsb.B == 0)
+                ld = 0.0;
+            else if (hsb.B == 0xff)
+                ld = 1.0;
+            else
+                ld = Math.Exp((hsb.B - 257.7) / 149.9);
 
             // HSLToRGB requires H, L and S to be in 0..1 range.
             hsl.H = (hsb.H) / 255.0;
@@ -80,139 +103,101 @@ namespace RibbonLib
             return hsl;
         }
 
-        // based on http://www.geekymonkey.com/Programming/CSharp/RGB2HSL_HSL2RGB.htm
-        // Given H,S,L in range of 0-1
-        // Returns a Color (RGB struct) in range of 0-255
-        public static Color HSLToRGB(HSL hsl)
+        //Conversion from W3C
+        //H [0..360], S [0..1], L [0..1]
+        static Color HslToRgb(double hue, double sat, double light)
         {
-            double v;
-            double r, g, b;
-
-            r = hsl.L;   // default to gray
-            g = hsl.L;
-            b = hsl.L;
-            v = (hsl.L <= 0.5) ? (hsl.L * (1.0 + hsl.S)) : (hsl.L + hsl.S - hsl.L * hsl.S);
-            if (v > 0)
+            double t1, t2;
+            int r, g, b;
+            hue = hue / 60.0;
+            if (light <= 0.5)
             {
-                double m;
-                double sv;
-                int sextant;
-                double fract, vsf, mid1, mid2;
-
-                m = hsl.L + hsl.L - v;
-                sv = (v - m) / v;
-                hsl.H *= 6.0;
-                sextant = (int)hsl.H;
-                fract = hsl.H - sextant;
-                vsf = v * sv * fract;
-                mid1 = m + vsf;
-                mid2 = v - vsf;
-
-                switch (sextant)
-                {
-                    case 0:
-                        r = v;
-                        g = mid1;
-                        b = m;
-                        break;
-
-                    case 1:
-                        r = mid2;
-                        g = v;
-                        b = m;
-                        break;
-
-                    case 2:
-                        r = m;
-                        g = v;
-                        b = mid1;
-                        break;
-
-                    case 3:
-                        r = m;
-                        g = mid2;
-                        b = v;
-                        break;
-
-                    case 4:
-                        r = mid1;
-                        g = m;
-                        b = v;
-                        break;
-
-                    case 5:
-                        r = v;
-                        g = m;
-                        b = mid2;
-                        break;
-                }
+                t2 = light * (sat + 1);
             }
-
-            return Color.FromArgb(Convert.ToByte(r * 255.0f), Convert.ToByte(g * 255.0f), Convert.ToByte(b * 255.0f));
+            else
+            {
+                t2 = light + sat - (light * sat);
+            }
+            t1 = light * 2 - t2;
+            r = Convert.ToByte(HueToRgb(t1, t2, hue + 2) * 255);
+            g = Convert.ToByte(HueToRgb(t1, t2, hue) * 255);
+            b = Convert.ToByte(HueToRgb(t1, t2, hue - 2) * 255);
+            return Color.FromArgb(r, g, b);
         }
 
-        // Given a Color (RGB Struct) in range of 0-255
-        // Return H,S,L in range of 0-1
-        public static HSL RGBToHSL(Color rgb)
+        static double HueToRgb(double t1, double t2, double hue)
         {
-            HSL hsl;
+            if (hue < 0) hue += 6;
+            if (hue >= 6) hue -= 6;
+            if (hue < 1) return (t2 - t1) * hue + t1;
+            else if (hue < 3) return t2;
+            else if (hue < 4) return (t2 - t1) * (4 - hue) + t1;
+            else return t1;
+        }
 
-            double r = rgb.R / 255.0;
-            double g = rgb.G / 255.0;
-            double b = rgb.B / 255.0;
-            double v;
-            double m;
-            double vm;
-            double r2, g2, b2;
-
-            hsl.H = 0; // default to black
-            hsl.S = 0;
-            hsl.L = 0;
-            v = Math.Max(r, g);
-            v = Math.Max(v, b);
-            m = Math.Min(r, g);
-            m = Math.Min(m, b);
-            hsl.L = (m + v) / 2.0;
-
-            if (hsl.L <= 0.0)
+        //Conversion from W3C
+        //H [0..360], S [0..1], L [0..1]
+        //same conversion like Microsoft .NET Color.Hue, Color.Saturation, Color.Brightness
+        static HSL RGBToHsl(byte r, byte g, byte b)
+        {
+            double min, max, l, s, maxcolor, h = 0;
+            int i;
+            double[] rgb = new double[3];
+            HSL hsl = new HSL();
+            rgb[0] = r / 255.0;
+            rgb[1] = g / 255.0;
+            rgb[2] = b / 255.0;
+            min = rgb[0];
+            max = rgb[0];
+            maxcolor = 0;
+            for (i = 0; i < rgb.Length - 1; i++)
             {
-                return hsl;
+                if (rgb[i + 1] <= min) { min = rgb[i + 1]; }
+                if (rgb[i + 1] >= max) { max = rgb[i + 1]; maxcolor = i + 1; }
             }
-
-            vm = v - m;
-            hsl.S = vm;
-
-            if (hsl.S > 0.0)
+            if (r == g && g == b)
+                h = 0;
+            else
             {
-                hsl.S /= (hsl.L <= 0.5) ? (v + m) : (2.0 - v - m);
+                if (maxcolor == 0)
+                {
+                    h = (rgb[1] - rgb[2]) / (max - min);
+                }
+                if (maxcolor == 1)
+                {
+                    h = 2 + (rgb[2] - rgb[0]) / (max - min);
+                }
+                if (maxcolor == 2)
+                {
+                    h = 4 + (rgb[0] - rgb[1]) / (max - min);
+                }
+
+                if (double.NaN == h) { h = 0; }
+                h = h * 60;
+                if (h < 0) { h = h + 360; }
+            }
+            l = (min + max) / 2;
+            if (min == max)
+            {
+                s = 0;
             }
             else
             {
-                return hsl;
+                if (l < 0.5)
+                {
+                    s = (max - min) / (max + min);
+                }
+                else
+                {
+                    s = (max - min) / (2 - max - min);
+                }
             }
-
-            r2 = (v - r) / vm;
-            g2 = (v - g) / vm;
-            b2 = (v - b) / vm;
-
-            if (r == v)
-            {
-                hsl.H = (g == m ? 5.0 + b2 : 1.0 - g2);
-            }
-            else if (g == v)
-            {
-                hsl.H = (b == m ? 1.0 + r2 : 3.0 - b2);
-            }
-            else
-            {
-                hsl.H = (r == m ? 3.0 + g2 : 5.0 - r2);
-            }
-
-            hsl.H /= 6.0;
-
+            hsl.H = h; hsl.S = s; hsl.L = l;
             return hsl;
         }
 
+        //HSL.H [0..1], HSL.S [0..1], HSL.L [0..1]
+        //HSB.H [0..255], HSB.S [0..255], HSB.L [0..255]
         public static HSB HSLToHSB(HSL hsl)
         {
             HSB hsb;
@@ -229,6 +214,7 @@ namespace RibbonLib
             return hsb;
         }
 
+        //HSB.H [0..255], HSB.S [0..255], HSB.L [0..255]
         public static uint HSBToUInt32(HSB hsb)
         {
             return (uint)(hsb.H | (hsb.S << 8) | (hsb.B << 16));